Fix 0 degrees pattern
The temperature started to be unreasonably high.
It turns out we couldn't find a match for the 0 on the y-axis.
And the assert didn't work as expected, so the script didn't fail.
Instead, it just returned a bogus value.

I found this great tool to inspect the pixels
of the picture in detail: https://pixspy.com