Change imageio.imread to imageio.v2.imread
Apparently, the function signature will change soon, so to
keep the current behavior, it's recommended to switch
to imageio.v2.imread.

The logged warning was:
DeprecationWarning: Starting with ImageIO v3 the behavior of this function will switch to that of iio.v3.imread. To keep the current behavior (and make this warning dissapear) use `import imageio.v2 as imageio` or call `imageio.v2.imread` directly.